packaging and documenting epiclomal
