+++ README.md
@@ -11,3 +11,65 @@ This folder contains the following:
 
 If you use this software, please cite "Epiclomal: probabilistic clustering of sparse single-cell DNA methylation data,
 Camila P. E. de Souza, Mirela Andronescu, Tehmina Masud, Farhia Kabeer, Justina Biele, Emma Laks, Daniel Lai, Jazmine Brimhall, Beixi Wang, Edmund Su, Tony Hui, Qi Cao, Marcus Wong, Michelle Moksa, Richard A. Moore, Martin Hirst, Samuel Aparicio, Sohrab P. Shah, doi: https://doi.org/10.1101/414482"
+
+
+## Setup and Installation
+
+Set up conda with the required packages.
+
+First ensure you have the correct channels:
+```
+conda config --add channels 'https://conda.anaconda.org/dranew'
+conda config --add channels 'https://conda.anaconda.org/aroth85'
+conda config --add channels 'https://conda.anaconda.org/shahcompbio'
+conda config --add channels 'bioconda'
+conda config --add channels 'r'
+conda config --add channels 'conda-forge'
+```
+
+### From Source
+
+Clone Epiclomal:
+
+```
+git clone https://github.com/shahcompbio/Epiclomal.git
+cd Epiclomal
+```
+
+Then create an environment with the required packages:
+
+```
+conda create --name Epiclomal --file conda_packages.txt
+```
+
+Activate the environment:
+
+```
+source activate Epiclomal
+```
+
+Add the ctDNA pipeline into the current site packages:
+```
+python setup.py install
+```
+
+## Usage
+
+### Run with generated synthetic data
+
+A Snakemake workflow exists to generate synthetic data and run the clustering and cluster evaluation software against the generated data.
+To run the Snakemake workflow, first edit the config file found at Epiclomal/snakemake/synthetic_data/config.yaml with appropriate paths and parameters. Then run
+```
+snakemake -s /path/to/Epiclomal/snakemake/synthetic_data/Snakefile
+```
+to run the workflow locally. To submit the jobs on the shahlab cluster and with parallelization, run
+```
+snakemake -s /path/to/Epiclomal/snakemake/synthetic_data/Snakefile --cluster 'qsub -V -hard -q shahlab.q -l h_vmem=8G -P shahlab_high -S /bin/bash' -j 32
+```
+
+To run each step of the synthetic data workflow individually, follow the steps outlined here: https://github.com/shahcompbio/Epiclomal/blob/master/examples/README.md
+
+## Input
+## Output
+
+
