+++ conda_packages.txt
@@ -3,76 +3,131 @@
 # platform: linux-64
 _libgcc_mutex=0.1=main
 _r-mutex=1.0.1=anacondar_1
+aioeasywebdav=2.4.0=py37_1000
+aiohttp=3.6.0=py37h516909a_0
+appdirs=1.4.3=py_1
+asn1crypto=0.24.0=py37_1003
+async-timeout=3.0.1=py_1000
+attrs=19.1.0=py_0
+bcrypt=3.1.6=py37h516909a_1
 binutils_impl_linux-64=2.31.1=h6176602_1
 binutils_linux-64=2.31.1=h6176602_8
 bioconductor-biobase=2.44.0=r36h516909a_1
 bioconductor-biocgenerics=0.30.0=r36_1
 bioconductor-pcamethods=1.76.0=r36he1b5a44_1
+boto3=1.9.231=py_0
+botocore=1.12.231=py_0
 bwidget=1.9.11=0
-bzip2=1.0.8=h516909a_0
-ca-certificates=2019.6.16=hecc5488_0
-cairo=1.16.0=hfb77d84_1002
-certifi=2019.6.16=py37_1
+bzip2=1.0.8=h516909a_1
+ca-certificates=2019.9.11=hecc5488_0
+cachetools=2.1.0=py_0
+cairo=1.16.0=h18b612c_1001
+certifi=2019.9.11=py37_0
+cffi=1.12.3=py37h8022711_0
+chardet=3.0.4=py37_1003
+configargparse=0.13.0=py_1
+cryptography=2.7=py37h72c5cf5_0
 curl=7.65.3=hf8cf82a_0
-fontconfig=2.13.1=h86ecdb6_1001
-freetype=2.10.0=he983fc9_0
-fribidi=1.0.5=h516909a_1002
+datrie=0.8=py37h516909a_0
+decorator=4.4.0=py_0
+docutils=0.15.2=py37_0
+dropbox=9.4.0=py_0
+epiclomal=0.0.1=dev_0
+expat=2.2.5=he1b5a44_1003
+filechunkio=1.8=py_2
+fontconfig=2.13.1=he4413a7_1000
+freetype=2.10.0=he983fc9_1
+ftputil=3.4=py_0
 gcc_impl_linux-64=7.3.0=habb00fd_1
 gcc_linux-64=7.3.0=h553295d_8
 gettext=0.19.8.1=hc5be6a0_1002
 gfortran_impl_linux-64=7.3.0=hdf63c60_1
 gfortran_linux-64=7.3.0=h553295d_8
+gitdb2=2.0.5=py_0
+gitpython=3.0.2=py_0
 glib=2.58.3=h6f030ca_1002
+google-api-core=1.14.2=py37_0
+google-auth=1.6.3=py_0
+google-cloud-core=1.0.3=py_0
+google-cloud-storage=1.19.0=py_0
+google-resumable-media=0.3.3=py_0
+googleapis-common-protos=1.6.0=py37_0
 graphite2=1.3.13=hf484d3e_1000
-gsl=2.5=h294904e_0
+graphviz=2.38.0=hf68f40c_1011
+gsl=2.5=h294904e_1
 gxx_impl_linux-64=7.3.0=hdf63c60_1
 gxx_linux-64=7.3.0=h553295d_8
-harfbuzz=2.4.0=h9f30f68_2
-icu=64.2=he1b5a44_0
+harfbuzz=2.4.0=h37c48d4_1
+icu=58.2=hf484d3e_1000
+idna=2.8=py37_1000
+jinja2=2.10.1=py_0
+jmespath=0.9.4=py_0
 joblib=0.13.2=py_0
 jpeg=9c=h14c3975_1001
+jsonschema=3.0.2=py37_0
 krb5=1.16.3=h05b26f9_1001
-libblas=3.8.0=11_openblas
-libcblas=3.8.0=11_openblas
+libblas=3.8.0=12_openblas
+libcblas=3.8.0=12_openblas
 libcurl=7.65.3=hda55be3_0
 libedit=3.1.20170329=hf8c457e_1001
 libffi=3.2.1=he1b5a44_1006
 libgcc-ng=9.1.0=hdf63c60_0
 libgfortran-ng=7.3.0=hdf63c60_0
 libiconv=1.15=h516909a_1005
-liblapack=3.8.0=11_openblas
-libopenblas=0.3.6=h6e990d7_6
+liblapack=3.8.0=12_openblas
+libopenblas=0.3.7=h6e990d7_1
 libpng=1.6.37=hed695b0_0
+libprotobuf=3.9.1=h8b12597_0
 libssh2=1.8.2=h22169c7_2
 libstdcxx-ng=9.1.0=hdf63c60_0
 libtiff=4.0.10=h57b8799_1003
+libtool=2.4.6=h14c3975_1002
 libuuid=2.32.1=h14c3975_1000
 libxcb=1.13=h14c3975_1002
-libxml2=2.9.9=hee79883_2
+libxml2=2.9.9=h13577e0_2
+llvmlite=0.29.0=py37hfd453ef_1
 lz4-c=1.8.3=he1b5a44_1001
 make=4.2.1=h14c3975_2004
+markupsafe=1.1.1=py37h14c3975_0
 memory_profiler=0.55.0=py_0
+multidict=4.5.2=py37h14c3975_1000
 ncurses=6.1=hf484d3e_1002
-numpy=1.17.0=py37h95a1406_0
+networkx=2.3=py_0
+numba=0.45.1=py37hb3f55d8_0
+numpy=1.17.2=py37h95a1406_0
 openssl=1.1.1c=h516909a_0
 pandas=0.25.0=py37hb3f55d8_0
-pango=1.42.4=he7ab937_0
+pango=1.40.14=he7ab937_1005
+paramiko=2.6.0=py37_0
 pcre=8.41=hf484d3e_1003
-pip=19.2.1=py37_0
+pip=19.2.3=py37_0
 pixman=0.38.0=h516909a_1003
+prettytable=0.7.2=py_3
+protobuf=3.9.1=py37he1b5a44_0
 psutil=5.6.3=py37h516909a_0
 pthread-stubs=0.4=h14c3975_1001
+pyasn1=0.4.6=py_0
+pyasn1-modules=0.2.6=py_0
+pycparser=2.19=py37_1
+pygments=2.4.2=py_0
+pygraphviz=1.5=py37h516909a_1001
+pynacl=1.3.0=py37h14c3975_1000
+pyopenssl=19.0.0=py37_0
+pyrsistent=0.15.4=py37h516909a_0
+pysftp=0.2.9=py_1
+pysocks=1.7.0=py37_0
 python=3.7.3=h33d41f4_1
 python-dateutil=2.8.0=py_0
+python-irodsclient=0.7.0=py_0
 pytz=2019.2=py_0
 pyyaml=5.1.2=py37h516909a_0
 r-ape=5.3=r36h0357c0b_1
 r-argparse=2.0.1=r36h6115d3f_0
-r-base=3.6.1=h6e652e1_3
+r-base=3.6.1=h8900bf8_2
 r-coda=0.19_3=r36h6115d3f_1
 r-colorspace=1.4_1=r36hcdcec82_1
 r-evaluate=0.14=r36h6115d3f_1
-r-findpython=1.0.5=r36h6115d3f_0
+r-findpython=1.0.5=r36h6115d3f_1
 r-fnn=1.1.3=r36h0357c0b_1
 r-glue=1.3.1=r36hcdcec82_1
 r-gtable=0.3.0=r36h6115d3f_2
@@ -82,7 +137,7 @@ r-knitr=1.23=r36h6115d3f_1
 r-labeling=0.3=r36h6115d3f_1002
 r-lattice=0.20_38=r36hcdcec82_1002
 r-magrittr=1.5=r36h6115d3f_1002
-r-markdown=1.0=r36hcdcec82_1
+r-markdown=1.1=r36hcdcec82_0
 r-mass=7.3_51.4=r36hcdcec82_1
 r-matrix=1.2_17=r36hcdcec82_1
 r-matrixmodels=0.4_1=r36h6115d3f_1002
@@ -92,30 +147,40 @@ r-mime=0.7=r36hcdcec82_1
 r-munsell=0.5.0=r36h6115d3f_1002
 r-mvtnorm=1.0_11=r36h9bbef5b_1
 r-nbclust=3.0=r36h6115d3f_1001
-r-nlme=3.1_140=r36h9bbef5b_1
+r-nlme=3.1_141=r36h9bbef5b_1
 r-pheatmap=1.0.12=r36h6115d3f_1
-r-quantreg=5.42.1=r36h9bbef5b_0
+r-quantreg=5.51=r36h9bbef5b_0
 r-r6=2.4.0=r36h6115d3f_2
 r-rcolorbrewer=1.1_2=r36h6115d3f_1002
 r-rcpp=1.0.2=r36h0357c0b_0
 r-rcppannoy=0.0.12=r36h0357c0b_1
 r-scales=1.0.0=r36h0357c0b_1002
-r-sparsem=1.77=r36h9bbef5b_1002
-r-stringi=1.4.3=r36h0e574ca_3
+r-sparsem=1.77=r36h9bbef5b_1003
+r-stringi=1.4.3=r36h0357c0b_2
 r-stringr=1.4.0=r36h6115d3f_1
 r-viridislite=0.3.0=r36h6115d3f_1002
-r-xfun=0.8=r36h6115d3f_1
+r-xfun=0.9=r36h6115d3f_0
 r-yaml=2.2.0=r36hcdcec82_1002
+ratelimiter=1.2.0=py37_1000
 readline=8.0=hf8c457e_0
+requests=2.22.0=py37_1
+rsa=3.4.2=py_1
+s3transfer=0.2.1=py37_0
 scikit-learn=0.21.3=py37hcdab131_0
-scipy=1.3.0=py37h921218d_1
+scipy=1.3.1=py37h921218d_2
 sed=4.7=h1bed415_1000
-setuptools=41.0.1=py37_0
+setuptools=41.2.0=py37_0
 six=1.12.0=py37_1000
-sqlite=3.29.0=hcee41ef_0
-tk=8.6.9=hed695b0_1002
-tktable=2.10=h555a92e_1
-wheel=0.33.4=py37_0
+smmap2=2.0.5=py_0
+snakemake=5.5.4=1
+snakemake-minimal=5.5.4=py_2
+sqlite=3.29.0=hcee41ef_1
+tk=8.6.9=hed695b0_1003
+tktable=2.10=h555a92e_2
+urllib3=1.25.3=py37_0
+wheel=0.33.6=py37_0
+wrapt=1.11.2=py37h516909a_0
+xmlrunner=1.7.7=py_0
 xorg-kbproto=1.0.7=h14c3975_1002
 xorg-libice=1.0.10=h516909a_0
 xorg-libsm=1.2.3=h84519dc_1000
@@ -123,11 +188,14 @@ xorg-libx11=1.6.8=h516909a_0
 xorg-libxau=1.0.9=h14c3975_0
 xorg-libxdmcp=1.1.3=h516909a_0
 xorg-libxext=1.3.4=h516909a_0
+xorg-libxpm=3.5.12=h516909a_1002
 xorg-libxrender=0.9.10=h516909a_1002
+xorg-libxt=1.1.5=h516909a_1003
 xorg-renderproto=0.11.1=h14c3975_1002
 xorg-xextproto=7.3.0=h14c3975_1002
 xorg-xproto=7.0.31=h14c3975_1007
 xz=5.2.4=h14c3975_1001
 yaml=0.1.7=h14c3975_1001
-zlib=1.2.11=h516909a_1005
+yarl=1.3.0=py37h14c3975_1000
+zlib=1.2.11=h516909a_1006
 zstd=1.4.0=h3b9ef0a_0
