+++ snakemake/synthetic_data/Snakefile
@@ -190,7 +190,7 @@ rule run_euclidean_eval:
     output:
         os.path.join(NPM_DIRECTORY, "results_EuclideanClust.txt")
     shell:
-        'python {config[SOFTWARE][evaluate_clustering_software]} '
+        'evaluate_clustering '
         '--true_clusters_file {input.true_clusters_file} '
         '--true_prevalences {config[PARAMS][ALL][prevalences]} '
         '--predicted_clusters_file {input.hfile} '
@@ -204,7 +204,7 @@ rule run_hamming_eval:
     output:
         os.path.join(NPM_DIRECTORY, "results_HammingClust.txt")
     shell:
-        'python {config[SOFTWARE][evaluate_clustering_software]} '
+        'evaluate_clustering '
         '--true_clusters_file {input.true_clusters_file} '
         '--true_prevalences {config[PARAMS][ALL][prevalences]} '
         '--predicted_clusters_file {input.pfile} '
@@ -218,7 +218,7 @@ rule run_pearson_eval:
     output:
         os.path.join(NPM_DIRECTORY, "results_PearsonClust.txt")
     shell:
-        'python {config[SOFTWARE][evaluate_clustering_software]} '
+        'evaluate_clustering '
         '--true_clusters_file {input.true_clusters_file} '
         '--true_prevalences {config[PARAMS][ALL][prevalences]} '
         '--predicted_clusters_file {input.peafile} '
@@ -232,7 +232,7 @@ rule run_densitycut_eval:
     output:
         os.path.join(NPM_DIRECTORY, "results_DensityCut.txt")
     shell:
-        'python {config[SOFTWARE][evaluate_clustering_software]} '
+        'evaluate_clustering '
         '--true_clusters_file {input.true_clusters_file} '
         '--true_prevalences {config[PARAMS][ALL][prevalences]} '
         '--predicted_clusters_file {input.dfile} '
@@ -252,7 +252,7 @@ rule run_epiclomal_basic:
         '(echo "==================================="; '
         'echo "Running EpiclomalBasic RUN {wildcards.run}"; '
         'echo "==================================="; '
-        'python ../../epiclomal/epiclomal.py Basic-GeMM '
+        'epiclomal Basic-GeMM '
         '--out_dir {output} '
         '--repeat_id {wildcards.run} '
         '--K {config[PARAMS][EPICLOMAL][K]} '
@@ -279,7 +279,7 @@ rule run_epiclomal_region:
         '(echo "==================================="; '
         'echo "Running EpiclomalRegion RUN {wildcards.run}"; '
         'echo "==================================="; '
-        'python ../../epiclomal/epiclomal.py Region-GeMM '
+        'epiclomal Region-GeMM '
         '--out_dir {output} '
         '--repeat_id {wildcards.run} '
         '--K {config[PARAMS][EPICLOMAL][K]} '
