+++ snakemake/synthetic_data/config.yaml
@@ -1,10 +1,10 @@
 OUTPUT_DIRS:
-  DATA_DIRECTORY: '../../examples/synthetic'
-  NPM_DIRECTORY: '../../examples/simple_hclust'
-  EPI_BASIC_DIR: '../../examples/epiclomal_basic/'
-  EPI_REGION_DIR: '../../examples/epiclomal_region/'
-  EVAL_BASIC_DIR: '../../examples/results_basic'
-  EVAL_REGION_DIR: '../../examples/results_region'
+  DATA_DIRECTORY: '/shahlab/pye/projects/epiclomal/Epiclomal/examples/synthetic'
+  NPM_DIRECTORY: '/shahlab/pye/projects/epiclomal/Epiclomal/examples/simple_hclust'
+  EPI_BASIC_DIR: '/shahlab/pye/projects/epiclomal/Epiclomal/examples/epiclomal_basic/'
+  EPI_REGION_DIR: '/shahlab/pye/projects/epiclomal/Epiclomal/examples/epiclomal_region/'
+  EVAL_BASIC_DIR: '/shahlab/pye/projects/epiclomal/Epiclomal/examples/results_basic'
+  EVAL_REGION_DIR: '/shahlab/pye/projects/epiclomal/Epiclomal/examples/results_region'
 
 LOG_DIR: 'log'
 
@@ -45,16 +45,15 @@ PARAMS:
     max_num_iters: 1000
     Bishop_model_selection: False
     check_uncertainty: True
-    config_file: '../../examples/inputs/config1.yaml'
+    config_file: '/shahlab/pye/projects/epiclomal/Epiclomal/examples/inputs/config1.yaml'
 
 SOFTWARE:
-  generate_synthetic_software: '../../scripts/generate_synthetic.R'
-  visualization_software: '../../scripts/visualization.R'
-  euclidean_clust_software: '../../scripts/non_probabilistic_methods/euclideanclust.R'
-  hamming_clust_software: '../../scripts/non_probabilistic_methods/hammingclust.R'
-  pearson_clust_software: '../../scripts/non_probabilistic_methods/pearsonclust.R'
-  density_cut_software: '../../scripts/non_probabilistic_methods/densitycut.R'
-  evaluate_clustering_software: '../../epiclomal/evaluate_clustering.py'
-  epiclomal_software: '../../epiclomal/epiclomal.py'
-  eval_software: '../../scripts/eval_epiclomal.R'
-  hdist_software: '../../scripts/hamming_distance.R'
+  generate_synthetic_software: '/shahlab/pye/projects/epiclomal/Epiclomal/scripts/generate_synthetic.R'
+  visualization_software: '/shahlab/pye/projects/epiclomal/Epiclomal/scripts/visualization.R'
+  euclidean_clust_software: '/shahlab/pye/projects/epiclomal/Epiclomal/scripts/non_probabilistic_methods/euclideanclust.R'
+  hamming_clust_software: '/shahlab/pye/projects/epiclomal/Epiclomal/scripts/non_probabilistic_methods/hammingclust.R'
+  pearson_clust_software: '/shahlab/pye/projects/epiclomal/Epiclomal/scripts/non_probabilistic_methods/pearsonclust.R'
+  density_cut_software: '/shahlab/pye/projects/epiclomal/Epiclomal/scripts/non_probabilistic_methods/densitycut.R'
+  epiclomal_software: '/shahlab/pye/projects/epiclomal/Epiclomal/epiclomal/epiclomal.py'
+  eval_software: '/shahlab/pye/projects/epiclomal/Epiclomal/scripts/eval_epiclomal.R'
+  hdist_software: '/shahlab/pye/projects/epiclomal/Epiclomal/scripts/hamming_distance.R'
